完善Object： 添加分类； 添加形状和点的约束关系。

+++ Format/Point.cs
@@ -9,9 +9,9 @@ namespace Vistart.ImageRecognitionAnnotationFormat.Format
     public class Point
     {
         [JsonProperty]
-        public int X;
+        public float X;
 
         [JsonProperty]
-        public int Y;
+        public float Y;
     }
 }
